pyproject.toml
[project]
name = "logs-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "kafka-python>=2.2.13",
]
